requirements.txt
appdirs==1.4.4
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
Django==3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
idna==2.10
isodate==0.6.0
lxml==4.6.3
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
zeep==3.4.0